Modification: get_workspace method moved to utils.routing
